move jwtpayload extension to authHelper
