+++ apps/frontend/auth.ts
@@ -5,7 +5,6 @@ import Credentials from "next-auth/providers/credentials";
 import { validateJWT } from "./lib/authHelpers";
 import { NextResponse } from "next/server";
 import { getOrCreateUser } from "./lib/api/user";
-import { JwtPayload } from "jsonwebtoken";
 
 type User = {
   apiId: string;
@@ -14,14 +13,6 @@ type User = {
   email: string;
   // Add other fields as needed
 };
-
-interface DynamicJwtPayload extends JwtPayload {
-  verified_credentials: {
-    address?: string
-    chain?: string
-    format: string
-  }[]
-}
 
 export const config = {
   providers: [
@@ -38,18 +29,18 @@ export const config = {
           throw new Error("Token is required");
         }
 
-        const jwtPayload = await validateJWT(token) as DynamicJwtPayload;
-        const blockChainCredentials = jwtPayload.verified_credentials.filter(credential =>
+        const dynamicJwtPayload = await validateJWT(token);
+        const blockChainCredentials = dynamicJwtPayload?.verified_credentials.filter(credential =>
           credential.format === 'blockchain'
         )
 
         console.debug("Credentials filtered for blockchain", blockChainCredentials)
 
         const apiUserPayload = {
-          public_key: blockChainCredentials.find(credential => 
+          public_key: blockChainCredentials?.find(credential => 
             credential.chain === 'eip155'
           )?.address ?? "",
-          public_key_sei: blockChainCredentials.find(credential => 
+          public_key_sei: blockChainCredentials?.find(credential => 
             credential.chain === 'cosmos'
           )?.address ?? "",
         }
@@ -59,14 +50,14 @@ export const config = {
           throw new Error("While examining JWT Payload, user was found to not have both an EVM and SEI address!!!")
         }
 
-        if (jwtPayload) {
+        if (dynamicJwtPayload) {
           // Transform the JWT payload into your user object
           const apiUser = await getOrCreateUser(apiUserPayload)
           const user: User = {
             apiId: apiUser.id,
-            id: jwtPayload.sub ?? "", // Assuming 'sub' is the user ID
-            name: jwtPayload.name ?? "", // Replace with actual field from JWT payload
-            email: jwtPayload.email ?? "", // Replace with actual field from JWT payload
+            id: dynamicJwtPayload.sub ?? "", // Assuming 'sub' is the user ID
+            name: dynamicJwtPayload.name ?? "", // Replace with actual field from JWT payload
+            email: dynamicJwtPayload.email ?? "", // Replace with actual field from JWT payload
             // Map other fields as needed
           };
           return user;
